Fixed BCM host library problem

+++ mailbox.c
@@ -28,6 +28,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
+#include <inttypes.h>
 #include <fcntl.h>
 #include <unistd.h>
 #include <assert.h>
@@ -61,7 +62,7 @@ void *mapmem(unsigned base, unsigned size)
    printf("base=0x%x, mem=%p\n", base, mem);
 #endif
    if (mem == MAP_FAILED) {
-      printf("mmap error %d\n", (int)mem);
+      printf("mmap error %" PRIdPTR "\n", (intptr_t)mem);
       exit (-1);
    }
    close(mem_fd);
